improve codes

+++ Mixer.h
@@ -7,6 +7,6 @@ class Mixer
 public:
   inline static int8_t clock(int8_t a, int8_t b, int8_t c)
   {
-    return ((int16_t) a + b + c) >> 2;
+    return (a + b + c) >> 2;
   }
 };
