+++ VCA.h
@@ -7,6 +7,6 @@ class VCA
 public:
   inline static int8_t clock(int8_t a, uint8_t k)
   {
-    return highByte(a * (k << 1));
+    return highByte(a * (uint8_t) (k << (uint8_t) 1));
   }
 };
